Move EdDSA private key handling to new library
