+++ apps/passport-server/.env.example
@@ -88,7 +88,7 @@ BYPASS_EMAIL_REGISTRATION=true
 # the project:
 # node -e 'console.log(Buffer.from(new (require("node-rsa"))({b:2048}).exportKey("private")).toString("base64"))' | pbcopy
 SERVER_RSA_PRIVATE_KEY_BASE64="LS0tLS1CRUdJTiBSU0EgUFJJVkFURSBLRVktLS0tLQpNSUlFcEFJQkFBS0NBUUVBanNGdkt0MGF4R0tBV1pTVUVKeWVNU21iTGpLYkpVTllucXpTcFNSZThPR1kvKzI1CmFHa01QcDlzb2ZhTGp3S0Myam14MmQ2ZkM4RDVVYVRDU3d5K1laL29pcUFTeXVob2kzTkVRbGUzS3Nwclcrd1MKS0xaL3N1K1Z3bm9US3NaLzRjNFo1K25CYUlhZjd2c2Vxd0diRHhNME92d3N0d3l4cHVzQUlvSzRPa25LU0ZMcApkdzlhMzZYVWNLUVNCK1NkRHc1a213UVI0YUllcVg3Yjk2Tk5sUjBQbGRZeWp6alNQSlZrTkQreklndWcwdkVMCk4xN1lSY3hjYlplK2JhK0RUN2ZkWWVHK3EwTkJxTXBFc0tYUm5mdzlieWlUWDdLc3dMaDBsOG01cFlTbjZZTTIKeHVDOHJURVZpa3F5dStMNGJXNjNQTzN2OE4rSUxIWm5QaDVsdndJREFRQUJBb0lCQURpaDFrY3I1ZVdJcWlCZgppQW03Q0NxNnRid3pDVWRVRzBEL2tLVkE4NFVlQ2tObGZJdEdPVHhTcEpid0p0YzZDTTBuM0QxeFNxa3dDZmh3Cm96K0FzN29vNU9peHJpTGU1aktFTE0wZzNTSlZvUmViWkd6UVVZS2I5MDVxZlZZWDN6Mm1OTExERWg0MFloZzkKZVNJZ0JmS0tETXRscFJ1TkpxRHdVMkpkZDFadnZCUHRyM1dmaEJyRG1iM0Z5dkg5OEFuS20xN1oxbFBoR2tUZgpENTM5Skt1YmZGcFV5TUhHY2MzRDBKOXhOY1B4MWFIVkQxUi9MbXBVcDRkTFJuVWIyVnVqdytUaEgyQ1JPMEhJCmJpQ2FzZElTa040Um1kcWJEK2N5NE8wS3FzcWJNL2JXZHNMU3M0MU9jT1ZzajBmQzVUbEdWdmpaNHk3ekFqSlEKL2ZqNm5Ca0NnWUVBMzBuSDNYelBWemcvRWIyOTNuQzRJeHJoZTFNY3JjdTlzWUpZdmxpOWs2ZGZzeDVIRytlKwpiR2NBYTF2WWp3S0p2dllRejZEV2hWbTN5bkk3bTFQKzQ0WEIrcHF1clJGbWRFYkF2eVJpQ0ExVnQrUTU2QWlDCkFJL2RzUXk2R1hGTEdtNXQ0S2Y2NWprbW02TVdaK2dhRUlRL2o0Q2VHdjVnd0UxTm5oeU5RdU1DZ1lFQW82dGEKSlNScHhDa2thTjEvK2tRaEZXbmRYQW1tN0lnZUlDYytmYXhTRDZvMTdDdEwwM0kwMWhBaUZyZVlqSWpBUTZSagpyckJCcC8yK2pXVFdnSjV5bmEvUkZZZll3NEJhbjN6MThCUER3akU1VXY2RUlXMlNkcCtGbVphQnVCNVlpejdICmFqYnl4YmxVdnlBbThYeU5lTzNPZUQwc3Y0L3JjaFhrOHpjcUhIVUNnWUVBZ3oxaWRCUi82Y3Z0aGdmNEtWNG4KWExyWC9SeUVwc0l4NUx1d1hYRGJQSVlNQWNPeUYvekdPQUR5aGllZHliQzhOYmw3ZUE0WkNKZ0xKcnF0dmtaSQpjb0tUMlRCQnF3ZW5LSURObVl0UTgxcWpIMHJObmt3aDNTS3lNdlpBQytSRE9HM2Z0MFFFZkh6cTdGK1hNOHVBClJtVDAxdXZLQWlPeHlsV2xZV0YyeGZFQ2dZQmZDTkl4R3ZHbkMxU3U1MTYyVFBVQTRCVytxWVFaZC9nMkoxTnIKMmJGVWt6cjRSVE9rTlg1a2hiMmxmem0rOEhEalZnTGpua2loYVA5S1RyRVd3ZW8va2FWUVlTOEdaZk01QkhBeAoxT2dvVERqQzdqSit6b2JyQmptbk12TmptNHRDWDFPZll3U1l0c0owc0JDc3krdkJ1aVBYUVg1SVJTTXJmWGNSCm1vaURPUUtCZ1FDNVZXN1l4MWUrRnE2VjVjUHkvWDJrK0FnNW95bjVESDVxNVVFbEllSjFPSFVOZVA5U0J3THAKZEgrNkxUMnFmVzZUaVpvVG9RZmVKTE42NTUwSXFQZlNrcW9VRWpobFUxbzZDaXdEaDh1MUdCdnlzZnZvTHlwTwpPdUIzbUNuMTdlTTVlZS9yN1h1Z21TZGw2bU1ZZ0VUNW5UM2sxL1lySEV1UXFHbDkvRHh2VkE9PQotLS0tLUVORCBSU0EgUFJJVkFURSBLRVktLS0tLQ=="
-# node -e 'console.log(require("@pcd/eddsa-pcd").newEdDSAPrivateKey())' | pbcopy
+# node -e 'console.log(require("@pcd/eddsa-crypto").newEdDSAPrivateKey())' | pbcopy
 SERVER_EDDSA_PRIVATE_KEY="860d116de20c1cb5bd7cb68fdb786da9d9fb35bf96c336d6dcaef64733701f20"
 
 # Account resets are strictly rate-limited. However, this can be turned off
@@ -129,7 +129,7 @@ GENERIC_ISSUANCE_CLIENT_URL="http://localhost:3005"
 # private key. You can generate one using the following command at the root of the project, 
 # after installing dependencies and building the project:
 # 
-# node -e 'console.log(require("@pcd/eddsa-pcd").newEdDSAPrivateKey())' | pbcopy
+# node -e 'console.log(require("@pcd/eddsa-crypto").newEdDSAPrivateKey())' | pbcopy
 GENERIC_ISSUANCE_EDDSA_PRIVATE_KEY="129806f9d68c3cf87450a968748fd31f01ba4e6f2078d80f24ebab23d98df7cf"
 
 # To recognise email PCD credentials, generic issuance needs to know the Zupass public key.
