+++ apps/passport-server/test/generic-issuance/pipelines/auto-issuance/autoIssuanceSpec.ts
@@ -1,5 +1,4 @@
-import { getEdDSAPublicKey } from "@pcd/eddsa-crypto";
-import { newEdDSAPrivateKey } from "@pcd/eddsa-pcd";
+import { getEdDSAPublicKey, newEdDSAPrivateKey } from "@pcd/eddsa-crypto";
 import { expectIsEdDSATicketPCD } from "@pcd/eddsa-ticket-pcd";
 import { EmailPCDPackage } from "@pcd/email-pcd";
 import {
