+++ packages/lib/eddsa-crypto/src/privateKey.ts
