+++ packages/lib/eddsa-crypto/src/publicKey.ts
