+++ packages/pcd/eddsa-pcd/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./EDDSAPCDPackage";
 export * from "./EdDSAPCD";
-export * from "./util/util";
